Combine discover and recipe's Recipe Detail in the same component
