--- client/app/discover/[id]/RecipeDetail.module.css
@@ -1,242 +1,0 @@
-.recipeWrapper {
-    margin-top: 3.5em;
-    margin-left: 8em;
-    display: flex;
-    height: 80vh;
-    background: linear-gradient(135deg, rgba(255,240,220,0.8) 0%, rgba(255,220,255,0.8) 100%);
-    min-height: 100vh;
-    margin: 0;
-    padding-top: 3.5em;
-    padding-left: 8em;
-}
-
-.recipeDetailsWrapper {
-    width: 40%;
-}
-
-.recipeTitle {
-    font-size: 2rem;
-    font-weight: 600;
-    margin-bottom: 1em;
-}
-
-.recipeImage {
-    width: 60%;
-    height: 40%;
-    border-radius: 10px;
-    object-fit: cover;
-    margin-bottom: 1em;
-}
-
-.recipeStatsWrapper {
-    display: flex;
-}
-
-.recipeTime {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: rgba(143, 195, 255, .73);
-    border-radius: 7px;
-    color: #506264;
-    font-size: .8em;
-    padding: .4em .8em;
-    margin-top: .5em;
-    margin-right: 1em;
-    width: auto;
-    font-weight: 500;
-}
-
-.recipeClock {
-    font-size: 1.4em !important;
-    margin-right: .4em
-}
-
-.recipeIngredients {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: rgba(227, 167, 255, .6);
-    border-radius: 7px;
-    color: #506264;
-    font-size: .8em;
-    padding: .4em .8em;
-    margin-top: .5em;
-    margin-right: 1em;
-    width: auto;
-    font-weight: 500;
-}
-
-.servingSize {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: rgba(255, 207, 150, .6);
-    border-radius: 7px;
-    color: #506264;
-    font-size: .8em;
-    padding: .4em .8em;
-    margin-top: .5em;
-    margin-right: 1em;
-    width: auto;
-    font-weight: 500;
-}
-
-.title {
-    font-size: 1.2em;
-    font-weight: 600;
-    margin-top: 1em;
-    margin-bottom: 1em;
-}
-
-.recipeInstructionWrapper {
-    width: 55%;
-}
-
-.ingredientWrapper {
-    height: 25%;
-    width: 100%;
-    column-count: 2;
-    column-gap: 40px;
-    column-fill: auto;
-}
-
-.ingredient {
-    color: #506264;
-    font-size: .9em;
-    font-weight: 500;
-    margin-bottom: .4em;
-}
-
-.instructionWrapper {
-    height: 40%;
-    overflow-y: auto;
-    overflow-x: hidden;
-}
-
-.instructionWrapper::-webkit-scrollbar {
-    width: 12px;
-}
-
-.instructionWrapper::-webkit-scrollbar-thumb {
-    background-color: rgba(156, 183, 210, .6);
-    border-radius: 10px;
-    border: 3px solid transparent;
-}
-
-.instructionWrapper::-webkit-scrollbar-thumb:hover {
-    background-color: rgba(156, 183, 210, 0.806);
-}
-
-.instructionWrapper::-webkit-scrollbar-track {
-    background-color: #ebeff2;
-    border-radius: 10px;
-}
-
-.instructionList {
-    color: #506264;
-    font-size: 1em;
-    font-weight: 500;
-    margin-bottom: .2em;
-    list-style: none;
-    counter-reset: custom-counter;
-}
-
-.instructionList li {
-    counter-increment: custom-counter;
-    position: relative;
-    margin-bottom: 1.3em;
-    display: flex;
-    align-items: center;
-}
-
-.instructionList li::before {
-    content: counter(custom-counter);
-    position: absolute;
-    color: #506264;
-    margin-left: -40px;
-    background-color: rgba(143, 195, 255, .73);
-    font-weight: 600;
-    font-size: .8em;
-    border-radius: 50%;
-    height: 1.9em;
-    width: 1.9em;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.madeButton {
-    background-color: #EC4A27;
-    color: white;
-    font-weight: 600;
-    font-size: 1.4em;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 25%;
-    height: 7%;
-    border-radius: 10px;
-    margin: 5% auto;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.madeButton:hover {
-    background-color: rgb(233, 69, 15);
-}
-
-.trackerWrapper {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    grid-template-rows: repeat(2, auto);
-    grid-row-gap: 2em;
-    grid-column-gap: .5em;
-    justify-items: center;
-    font-size: .8em;
-    font-weight: 600;
-    color: #506264;
-    width: 65%;
-    margin-top: 3em;
-    margin-left: -1.5em;
-}
-
-.trackerItem,
-.trackerVisual {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.trackerItemTitle {
-    margin-bottom: .5em;
-}
-
-.data {
-    margin-top: .9em;
-}
-
-.titleWrapper {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 90%;
-}
-
-.titleButtons {
-    color: #EC4A27;
-    cursor: pointer;
-}
-
-.titleButtons:hover {
-    color: rgb(233, 69, 15);
-}
-
-.button {
-    font-size: 2em !important;
-    margin-left: .5em;
-}
-
-.madeButtonText {
-    margin: 0.5em 1em;
-}
